refactor: 스크린 네이밍 변경
