+++ app/src/main/java/com/sopt/now/compose/feature/main/Screen.kt
@@ -5,5 +5,5 @@ sealed class Screen(val route: String) {
 
     data object SignUp : Screen(route = "signUp")
 
-    data object Home : Screen(route = "home")
+    data object MyPage : Screen(route = "myPage")
 }
